All components integrated into config
