+++ rompy/swan/config.py
@@ -7,7 +7,7 @@ from pydantic import field_validator, Field, model_validator
 from rompy.core import BaseConfig, Coordinate, RompyBaseModel, Spectrum, TimeRange
 from rompy.swan.boundary import DataBoundary
 from rompy.swan.components import boundary, cgrid, inpgrid, physics, numerics, startup, group
-from rompy.swan.components.group import STARTUP, PHYSICS, OUTPUT
+from rompy.swan.components.group import STARTUP, PHYSICS, OUTPUT, LOCKUP
 
 from .data import SwanDataGrid
 from .grid import SwanGrid
@@ -254,6 +254,7 @@ PHYSICS_TYPE = Annotated[PHYSICS, Field(description="Physics components")]
 PROP_TYPE = Annotated[numerics.PROP, Field(description="Propagation components")]
 NUMERIC_TYPE = Annotated[numerics.NUMERIC, Field(description="Numerics components")]
 OUTPUT_TYPE = Annotated[OUTPUT, Field(description="Output components")]
+LOCKUP_TYPE = Annotated[LOCKUP, Field(description="Output components")]
 CGRID_TYPES = Annotated[
     Union[cgrid.REGULAR, cgrid.CURVILINEAR, cgrid.UNSTRUCTURED],
     Field(description="Cgrid component", discriminator="model_type"),
@@ -293,6 +294,7 @@ class SwanConfigComponents(BaseConfig):
     prop: Optional[PROP_TYPE] = Field(default=None)
     numeric: Optional[NUMERIC_TYPE] = Field(default=None)
     output: Optional[OUTPUT_TYPE] = Field(default=None)
+    lockup: Optional[LOCKUP_TYPE] = Field(default=None)
 
     @model_validator(mode="after")
     def no_nor_if_spherical(self) -> "SwanConfigComponents":
