+++ rompy/templates/swancomp/{{runtime.run_id}}/INPUT
@@ -35,3 +35,8 @@
 !----------------------------------------------------------- Output -------------------------------------------------------------
 
 {% if config.output %}{{config.output.render()}}{% endif %}
+
+
+!----------------------------------------------------------- Lockup -------------------------------------------------------------
+
+{% if config.lockup %}{{config.lockup.render()}}{% endif %}
